Refactor CLI messaging
